+++ components/header.js
@@ -2,6 +2,7 @@ import '../pages/index.scss' // these are done to avoid a nasty ordering issue i
 import './typography.scss'
 import './images.scss'
 import './print.scss'
+import './links.scss'
 
 import css from './header.scss'
 import Links from './links'